win-Scripts_CVE_Shodan+nmap/requirements.txt
requests
pymupdf
shodan
playwright
sumy
nltk
python-nmap
deep-translator
beautifulsoup4